split crate structure into configs and traits
