+++ Cargo.toml
@@ -38,6 +38,7 @@ members = [
     "crates/rpc/rpc-types/",
     "crates/rpc/rpc-types-compat/",
     "crates/node-builder/",
+    "crates/node-api/",
     "crates/snapshot/",
     "crates/stages/",
     "crates/storage/codecs/",
@@ -115,6 +116,7 @@ reth-db = { path = "crates/storage/db" }
 reth-discv4 = { path = "crates/net/discv4" }
 reth-dns-discovery = { path = "crates/net/dns" }
 reth-node-builder = { path = "crates/node-builder" }
+reth-node-api = { path = "crates/node-api" }
 reth-downloaders = { path = "crates/net/downloaders" }
 reth-ecies = { path = "crates/net/ecies" }
 reth-eth-wire = { path = "crates/net/eth-wire" }
