+++ crates/node-builder/src/lib.rs
@@ -1,4 +1,4 @@
-//! Standalone crate for Reth configuration traits and builder types.
+//! Standalone crate for Reth configuration and builder types.
 
 #![doc(
     html_logo_url = "https://raw.githubusercontent.com/paradigmxyz/reth/main/assets/reth-docs.png",
@@ -7,10 +7,5 @@
 )]
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
-/// Contains the [EngineTypes] trait, used to configure the types used by the engine.
-pub mod engine_types;
-pub use engine_types::EngineTypes;
-
-/// Contains traits and types used to abstract over payload attributes types.
-pub mod payload_attributes;
-pub use payload_attributes::{PayloadAttributesTrait, PayloadBuilderAttributesTrait};
+pub mod engine;
+pub use engine::EthEngineTypes;
